+++ error.go
@@ -11,8 +11,14 @@ import (
 
 	"github.com/pkg/errors"
 
-	"go.elastic.co/apm/model"
 	"go.elastic.co/apm/stacktrace"
+)
+
+type errorKind int
+
+const (
+	errorKindException errorKind = iota
+	errorKindLog
 )
 
 // Recovered creates an Error with t.NewError(err), where
@@ -63,13 +69,9 @@ func (t *Tracer) NewError(err error) *Error {
 	if err == nil {
 		panic("NewError must be called with a non-nil error")
 	}
-	e := t.newError()
+	e := t.newError(errorKindException)
 	rand.Read(e.ID[:]) // ignore error, can't do anything about it
-	e.model.Exception.Message = truncateString(err.Error())
-	if e.model.Exception.Message == "" {
-		e.model.Exception.Message = "[EMPTY]"
-	}
-	initException(&e.model.Exception, errors.Cause(err))
+	initException(&e.exception, err)
 	initStacktrace(e, err)
 	if len(e.stacktrace) == 0 {
 		e.SetStacktrace(2)
@@ -84,40 +86,50 @@ func (t *Tracer) NewError(err error) *Error {
 //
 // If r.Message is empty, "[EMPTY]" will be used.
 func (t *Tracer) NewErrorLog(r ErrorLogRecord) *Error {
-	e := t.newError()
-	e.model.Log = model.Log{
-		Message:      truncateString(r.Message),
-		Level:        truncateString(r.Level),
-		LoggerName:   truncateString(r.LoggerName),
-		ParamMessage: truncateString(r.MessageFormat),
+	e := t.newError(errorKindLog)
+	e.log = ErrorLogRecord{
+		Message:       truncateString(r.Message),
+		MessageFormat: truncateString(r.MessageFormat),
+		Level:         truncateString(r.Level),
+		LoggerName:    truncateString(r.LoggerName),
 	}
-	if e.model.Log.Message == "" {
-		e.model.Log.Message = "[EMPTY]"
+	if e.log.Message == "" {
+		e.log.Message = "[EMPTY]"
 	}
 	return e
 }
 
 // newError returns a new Error associated with the Tracer.
-func (t *Tracer) newError() *Error {
-	e, _ := t.errorPool.Get().(*Error)
+func (t *Tracer) newError(kind errorKind) *Error {
+	e, _ := t.errorDataPool.Get().(*ErrorData)
 	if e == nil {
-		e = &Error{
+		e = &ErrorData{
 			tracer: t,
 			Context: Context{
 				captureBodyMask: CaptureBodyErrors,
 			},
 		}
 	}
+	e.kind = kind
 	e.Timestamp = time.Now()
-	return e
+	return &Error{ErrorData: e}
 }
 
 // Error describes an error occurring in the monitored service.
 type Error struct {
-	model           model.Error
-	tracer          *Tracer
-	stacktrace      []stacktrace.Frame
-	modelStacktrace []model.StacktraceFrame
+	// ErrorData holds the error data. This field is set to nil when
+	// the error's Send method is called.
+	*ErrorData
+}
+
+// ErrorData holds the details for an error, and is embedded inside Error.
+// When the error is sent, its ErrorData field will be set to nil.
+type ErrorData struct {
+	tracer     *Tracer
+	stacktrace []stacktrace.Frame
+	kind       errorKind
+	exception  exceptionData
+	log        ErrorLogRecord
 
 	// ID is the unique identifier of the error. This is set by
 	// the various error constructors, and is exposed only so
@@ -163,42 +175,56 @@ type Error struct {
 
 // SetTransaction sets TraceID, TransactionID, and ParentID to the transaction's IDs.
 //
-// This must be called before tx.End(). After SetTransaction returns, e may be sent
-// and tx ended in either order.
+// SetTransaction has no effect if called with an ended transaction.
 func (e *Error) SetTransaction(tx *Transaction) {
-	e.TraceID = tx.traceContext.Trace
-	e.ParentID = tx.traceContext.Span
+	tx.mu.RLock()
+	if !tx.ended() {
+		e.setTransactionData(tx.TransactionData)
+	}
+	tx.mu.RUnlock()
+}
+
+func (e *Error) setTransactionData(td *TransactionData) {
+	e.TraceID = td.traceContext.Trace
+	e.ParentID = td.traceContext.Span
 	e.TransactionID = e.ParentID
 }
 
 // SetSpan sets TraceID, TransactionID, and ParentID to the span's IDs. If you call
 // this, it is not necessary to call SetTransaction.
 //
-// This must be called before s.End(). After SetSpanreturns, e may be sent and e ended
-// in either order.
+// SetSpan has no effect if called with an ended span.
 func (e *Error) SetSpan(s *Span) {
+	s.mu.RLock()
+	if !s.ended() {
+		e.setSpanData(s.SpanData)
+	}
+	s.mu.RUnlock()
+}
+
+func (e *Error) setSpanData(s *SpanData) {
 	e.TraceID = s.traceContext.Trace
 	e.ParentID = s.traceContext.Span
 	e.TransactionID = s.transactionID
 }
 
-func (e *Error) reset() {
-	*e = Error{
-		tracer:          e.tracer,
-		stacktrace:      e.stacktrace[:0],
-		modelStacktrace: e.modelStacktrace[:0],
-		Context:         e.Context,
-	}
-	e.Context.reset()
-	e.tracer.errorPool.Put(e)
-}
-
 // Send enqueues the error for sending to the Elastic APM server.
-// The Error must not be used after this.
+//
+// Send will set e.ErrorData to nil, so the error must not be
+// modified after Send returns.
 func (e *Error) Send() {
-	if e == nil {
+	if e == nil || e.sent() {
 		return
 	}
+	e.ErrorData.enqueue()
+	e.ErrorData = nil
+}
+
+func (e *Error) sent() bool {
+	return e.ErrorData == nil
+}
+
+func (e *ErrorData) enqueue() {
 	select {
 	case e.tracer.errors <- e:
 	default:
@@ -210,71 +236,79 @@ func (e *Error) Send() {
 	}
 }
 
-func (e *Error) setStacktrace() {
-	if len(e.stacktrace) == 0 {
-		return
+func (e *ErrorData) reset() {
+	*e = ErrorData{
+		tracer:     e.tracer,
+		stacktrace: e.stacktrace[:0],
+		Context:    e.Context,
+		exception:  e.exception,
 	}
-	e.modelStacktrace = appendModelStacktraceFrames(e.modelStacktrace, e.stacktrace)
-	e.model.Log.Stacktrace = e.modelStacktrace
-	e.model.Exception.Stacktrace = e.modelStacktrace
+	e.Context.reset()
+	e.exception.reset()
+	e.tracer.errorDataPool.Put(e)
 }
 
-func (e *Error) setCulprit() {
-	if e.Culprit != "" {
-		e.model.Culprit = e.Culprit
-	} else if e.modelStacktrace != nil {
-		e.model.Culprit = stacktraceCulprit(e.modelStacktrace)
+type exceptionData struct {
+	message         string
+	attrs           map[string]interface{}
+	typeName        string
+	typePackagePath string
+	codeNumber      float64
+	codeString      string
+}
+
+func (e *exceptionData) reset() {
+	*e = exceptionData{
+		attrs: e.attrs,
+	}
+	for k := range e.attrs {
+		delete(e.attrs, k)
 	}
 }
 
-// stacktraceCulprit returns the first non-library stacktrace frame's
-// function name.
-func stacktraceCulprit(frames []model.StacktraceFrame) string {
-	for _, frame := range frames {
-		if !frame.LibraryFrame {
-			return frame.Function
-		}
-	}
-	return ""
-}
-
-func initException(e *model.Exception, err error) {
+func initException(e *exceptionData, err error) {
 	setAttr := func(k string, v interface{}) {
-		if e.Attributes == nil {
-			e.Attributes = make(map[string]interface{})
+		if e.attrs == nil {
+			e.attrs = make(map[string]interface{})
 		}
-		e.Attributes[k] = v
+		e.attrs[k] = v
 	}
+
+	e.message = truncateString(err.Error())
+	if e.message == "" {
+		e.message = "[EMPTY]"
+	}
+	err = errors.Cause(err)
 
 	// Set Module, Type, Attributes, and Code.
 	switch err := err.(type) {
 	case *net.OpError:
-		e.Module, e.Type = "net", "OpError"
+		e.typePackagePath, e.typeName = "net", "OpError"
 		setAttr("op", err.Op)
 		setAttr("net", err.Net)
 		setAttr("source", err.Source)
 		setAttr("addr", err.Addr)
 	case *os.LinkError:
-		e.Module, e.Type = "os", "LinkError"
+		e.typePackagePath, e.typeName = "os", "LinkError"
 		setAttr("op", err.Op)
 		setAttr("old", err.Old)
 		setAttr("new", err.New)
 	case *os.PathError:
-		e.Module, e.Type = "os", "PathError"
+		e.typePackagePath, e.typeName = "os", "PathError"
 		setAttr("op", err.Op)
 		setAttr("path", err.Path)
 	case *os.SyscallError:
-		e.Module, e.Type = "os", "SyscallError"
+		e.typePackagePath, e.typeName = "os", "SyscallError"
 		setAttr("syscall", err.Syscall)
 	case syscall.Errno:
-		e.Module, e.Type = "syscall", "Errno"
-		e.Code.Number = float64(uintptr(err))
+		e.typePackagePath, e.typeName = "syscall", "Errno"
+		e.codeNumber = float64(uintptr(err))
 	default:
 		t := reflect.TypeOf(err)
 		if t.Name() == "" && t.Kind() == reflect.Ptr {
 			t = t.Elem()
 		}
-		e.Module, e.Type = t.PkgPath(), t.Name()
+		e.typePackagePath, e.typeName = t.PkgPath(), t.Name()
 
 		// If the error implements Type, use that to
 		// override the type name determined through
@@ -282,7 +316,7 @@ func initException(e *model.Exception, err error) {
 		if err, ok := err.(interface {
 			Type() string
 		}); ok {
-			e.Type = err.Type()
+			e.typeName = err.Type()
 		}
 
 		// If the error implements a Code method, use
@@ -291,11 +325,11 @@ func initException(e *model.Exception, err error) {
 		case interface {
 			Code() string
 		}:
-			e.Code.String = err.Code()
+			e.codeString = err.Code()
 		case interface {
 			Code() float64
 		}:
-			e.Code.Number = err.Code()
+			e.codeNumber = err.Code()
 		}
 	}
 	if errTemporary(err) {
@@ -304,8 +338,8 @@ func initException(e *model.Exception, err error) {
 	if errTimeout(err) {
 		setAttr("timeout", true)
 	}
-	e.Code.String = truncateString(e.Code.String)
-	e.Type = truncateString(e.Type)
+	e.codeString = truncateString(e.codeString)
+	e.typeName = truncateString(e.typeName)
 }
 
 func initStacktrace(e *Error, err error) {