+++ internal/apmhostutil/docker_nonlinux.go
@@ -1,0 +1,13 @@
+// +build !linux
+
+package apmhostutil
+
+import (
+	"runtime"
+
+	"github.com/pkg/errors"
+)
+
+func dockerContainerID() (string, error) {
+	return "", errors.Errorf("docker container ID computation not implemented for %s", runtime.GOOS)
+}