+++ tracer.go
@@ -53,6 +53,7 @@ type options struct {
 	maxSpans              int
 	requestSize           int
 	bufferSize            int
+	metricsBufferSize     int
 	sampler               Sampler
 	sanitizedFieldNames   wildcard.Matchers
 	captureBody           CaptureBodyMode
@@ -96,6 +97,12 @@ func (opts *options) init(continueOnError bool) error {
 		errs = append(errs, err)
 	}
 
+	metricsBufferSize, err := initialMetricsBufferSize()
+	if err != nil {
+		metricsBufferSize = int(defaultMetricsBufferSize)
+		errs = append(errs, err)
+	}
+
 	maxSpans, err := initialMaxSpans()
 	if failed(err) {
 		maxSpans = defaultMaxSpans
@@ -132,6 +139,7 @@ func (opts *options) init(continueOnError bool) error {
 	opts.metricsInterval = metricsInterval
 	opts.requestSize = requestSize
 	opts.bufferSize = bufferSize
+	opts.metricsBufferSize = metricsBufferSize
 	opts.maxSpans = maxSpans
 	opts.sampler = sampler
 	opts.sanitizedFieldNames = initialSanitizedFieldNames()
@@ -172,16 +180,17 @@ type Tracer struct {
 	process *model.Process
 	system  *model.System
 
-	active           bool
-	bufferSize       int
-	closing          chan struct{}
-	closed           chan struct{}
-	forceFlush       chan chan<- struct{}
-	forceSendMetrics chan chan<- struct{}
-	configCommands   chan tracerConfigCommand
-	transactions     chan *Transaction
-	spans            chan *Span
-	errors           chan *Error
+	active            bool
+	bufferSize        int
+	metricsBufferSize int
+	closing           chan struct{}
+	closed            chan struct{}
+	forceFlush        chan chan<- struct{}
+	forceSendMetrics  chan chan<- struct{}
+	configCommands    chan tracerConfigCommand
+	transactions      chan *TransactionData
+	spans             chan *SpanData
+	errors            chan *ErrorData
 
 	statsMu sync.Mutex
 	stats   TracerStats
@@ -198,9 +207,9 @@ type Tracer struct {
 	captureBodyMu sync.RWMutex
 	captureBody   CaptureBodyMode
 
-	errorPool       sync.Pool
-	spanPool        sync.Pool
-	transactionPool sync.Pool
+	errorDataPool       sync.Pool
+	spanDataPool        sync.Pool
+	transactionDataPool sync.Pool
 }
 
 // NewTracer returns a new Tracer, using the default transport,
@@ -235,15 +244,16 @@ func newTracer(opts options) *Tracer {
 		forceFlush:            make(chan chan<- struct{}),
 		forceSendMetrics:      make(chan chan<- struct{}),
 		configCommands:        make(chan tracerConfigCommand),
-		transactions:          make(chan *Transaction, transactionsChannelCap),
-		spans:                 make(chan *Span, spansChannelCap),
-		errors:                make(chan *Error, errorsChannelCap),
+		transactions:          make(chan *TransactionData, transactionsChannelCap),
+		spans:                 make(chan *SpanData, spansChannelCap),
+		errors:                make(chan *ErrorData, errorsChannelCap),
 		maxSpans:              opts.maxSpans,
 		sampler:               opts.sampler,
 		captureBody:           opts.captureBody,
 		spanFramesMinDuration: opts.spanFramesMinDuration,
 		active:                opts.active,
 		bufferSize:            opts.bufferSize,
+		metricsBufferSize:     opts.metricsBufferSize,
 	}
 	t.Service.Name = opts.serviceName
 	t.Service.Version = opts.serviceVersion
@@ -477,7 +487,7 @@ func (t *Tracer) loop() {
 	var metadata []byte
 	var gracePeriod time.Duration = -1
 	var flushed chan<- struct{}
-	var requestBufTransactions, requestBufSpans, requestBufErrors, requestBufMetrics uint64
+	var requestBufTransactions, requestBufSpans, requestBufErrors, requestBufMetricsets uint64
 	zlibWriter, _ := zlib.NewWriterLevel(&requestBuf, zlib.BestSpeed)
 	zlibFlushed := true
 	zlibClosed := false
@@ -510,17 +520,18 @@ func (t *Tracer) loop() {
 		}
 	}()
 
+	var stats TracerStats
 	var metrics Metrics
 	var sentMetrics chan<- struct{}
 	var gatheringMetrics bool
+	var metricsTimerStart time.Time
+	metricsBuffer := ringbuffer.New(t.metricsBufferSize)
 	gatheredMetrics := make(chan struct{}, 1)
 	metricsTimer := time.NewTimer(0)
-	metricsTimerActive := false
 	if !metricsTimer.Stop() {
 		<-metricsTimer.C
 	}
 
-	var stats TracerStats
 	var cfg tracerConfig
 	buffer := ringbuffer.New(t.bufferSize)
 	buffer.Evicted = func(h ringbuffer.BlockHeader) {
@@ -534,23 +545,43 @@ func (t *Tracer) loop() {
 		}
 	}
 	modelWriter := modelWriter{
-		buffer: buffer,
-		cfg:    &cfg,
-		stats:  &stats,
+		buffer:        buffer,
+		metricsBuffer: metricsBuffer,
+		cfg:           &cfg,
+		stats:         &stats,
 	}
 
 	for {
 		var gatherMetrics bool
 		select {
 		case <-t.closing:
+			cancelContext() // informs transport that EOF is expected
 			iochanReader.CloseRead(io.EOF)
 			return
 		case cmd := <-t.configCommands:
 			oldMetricsInterval := cfg.metricsInterval
 			cmd(&cfg)
-			if !gatheringMetrics && oldMetricsInterval <= 0 && cfg.metricsInterval > 0 {
-				metricsTimer.Reset(cfg.metricsInterval)
-				metricsTimerActive = true
+			if !gatheringMetrics && cfg.metricsInterval != oldMetricsInterval {
+				if metricsTimerStart.IsZero() {
+					if cfg.metricsInterval > 0 {
+						metricsTimer.Reset(cfg.metricsInterval)
+						metricsTimerStart = time.Now()
+					}
+				} else {
+					if cfg.metricsInterval <= 0 {
+						metricsTimerStart = time.Time{}
+						if !metricsTimer.Stop() {
+							<-metricsTimer.C
+						}
+					} else {
+						alreadyPassed := time.Since(metricsTimerStart)
+						if alreadyPassed >= cfg.metricsInterval {
+							metricsTimer.Reset(0)
+						} else {
+							metricsTimer.Reset(cfg.metricsInterval - alreadyPassed)
+						}
+					}
+				}
 			}
 			continue
 		case tx := <-t.transactions:
@@ -565,22 +596,22 @@ func (t *Tracer) loop() {
 			requestTimerActive = false
 			closeRequest = true
 		case <-metricsTimer.C:
-			metricsTimerActive = false
+			metricsTimerStart = time.Time{}
 			gatherMetrics = !gatheringMetrics
 		case sentMetrics = <-t.forceSendMetrics:
-			if metricsTimerActive {
+			if !metricsTimerStart.IsZero() {
 				if !metricsTimer.Stop() {
 					<-metricsTimer.C
 				}
-				metricsTimerActive = false
+				metricsTimerStart = time.Time{}
 			}
 			gatherMetrics = !gatheringMetrics
 		case <-gatheredMetrics:
 			modelWriter.writeMetrics(&metrics)
 			gatheringMetrics = false
-			closeRequest = true
+			flushRequest = true
 			if cfg.metricsInterval > 0 {
-				metricsTimerActive = true
+				metricsTimerStart = time.Now()
 				metricsTimer.Reset(cfg.metricsInterval)
 			}
 		case flushed = <-t.forceFlush:
@@ -594,7 +625,7 @@ func (t *Tracer) loop() {
 			for n := len(t.errors); n > 0; n-- {
 				modelWriter.writeError(<-t.errors)
 			}
-			if !requestActive && buffer.Len() == 0 {
+			if !requestActive && buffer.Len() == 0 && metricsBuffer.Len() == 0 {
 				flushed <- struct{}{}
 				continue
 			}
@@ -626,11 +657,11 @@ func (t *Tracer) loop() {
 						return ""
 					}
 					cfg.logger.Debugf(
-						"sent request with %d transaction%s, %d span%s, %d error%s, %d metric%s",
+						"sent request with %d transaction%s, %d span%s, %d error%s, %d metricset%s",
 						requestBufTransactions, s(requestBufTransactions),
 						requestBufSpans, s(requestBufSpans),
 						requestBufErrors, s(requestBufErrors),
-						requestBufMetrics, s(requestBufMetrics),
+						requestBufMetricsets, s(requestBufMetricsets),
 					)
 				}
 			}
@@ -640,7 +671,7 @@ func (t *Tracer) loop() {
 				t.statsMu.Unlock()
 				stats = TracerStats{}
 			}
-			if sentMetrics != nil {
+			if sentMetrics != nil && requestBufMetricsets > 0 {
 				sentMetrics <- struct{}{}
 				sentMetrics = nil
 			}
@@ -662,7 +693,7 @@ func (t *Tracer) loop() {
 			requestBufTransactions = 0
 			requestBufSpans = 0
 			requestBufErrors = 0
-			requestBufMetrics = 0
+			requestBufMetricsets = 0
 			if requestTimerActive {
 				if !requestTimer.Stop() {
 					<-requestTimer.C
@@ -687,7 +718,7 @@ func (t *Tracer) loop() {
 		}
 
 		if !requestActive {
-			if buffer.Len() == 0 {
+			if buffer.Len() == 0 && metricsBuffer.Len() == 0 {
 				continue
 			}
 			sendStreamRequest <- gracePeriod
@@ -704,7 +735,24 @@ func (t *Tracer) loop() {
 		}
 
 		if !closeRequest || !zlibClosed {
-			for requestBytesRead+requestBuf.Len() < cfg.requestSize && buffer.Len() > 0 {
+			for requestBytesRead+requestBuf.Len() < cfg.requestSize {
+				if metricsBuffer.Len() > 0 {
+					if _, _, err := metricsBuffer.WriteBlockTo(zlibWriter); err == nil {
+						requestBufMetricsets++
+						zlibWriter.Write([]byte("\n"))
+						zlibFlushed = false
+						if sentMetrics != nil {
+							// SendMetrics was called: close the request
+							// off so we can inform the user when the
+							// metrics have been processed.
+							closeRequest = true
+						}
+					}
+					continue
+				}
+				if buffer.Len() == 0 {
+					break
+				}
 				if h, _, err := buffer.WriteBlockTo(zlibWriter); err == nil {
 					switch h.Tag {
 					case transactionBlockTag:
@@ -713,8 +761,6 @@ func (t *Tracer) loop() {
 						requestBufSpans++
 					case errorBlockTag:
 						requestBufErrors++
-					case metricsBlockTag:
-						requestBufMetrics++
 					}
 					zlibWriter.Write([]byte("\n"))
 					zlibFlushed = false