+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased](https://github.com/elastic/apm-agent-go/compare/v1.0.0...master)
 
+ - Stop pooling Transaction/Span/Error, introduce internal pooled objects (#319)
+ - Enable metrics collection with default interval of 30s (#322)
+ - `ELASTIC_APM_SERVER_CERT` enables server certificate pinning (#325)
+ - Add Docker container ID to metadata (#330)
+
 ## [v1.0.0](https://github.com/elastic/apm-agent-go/releases/tag/v1.0.0)
 
  - Implement v2 intake protocol (#180)